requirements.txt
Flask
gunicorn
tensorflow-cpu==2.15.0
numpy
opencv-python-headless